Switching to update for lobbyplayer.
When using update against dynamodb it means it'll be created if the record doesn't exist or updated if it does. This is perfect for us as it then means we can do one call to create / update and retrieve the data out rather than throwing if statements into the code.